new server based on a storage
